requirements.txt
langchain_community
python-docx